tests: make tests run in CI on windows
They were failing due to missing dll. This commit
- inject the location of Qt dll in the PATH
- Qt plugin directory in QT_PLUGIN_PATH
- Qml import path in QML_IMPORT_PATH & QML2_IMPORT_PATH

They are not necessary to every tests, but at least they should be robust for the future when more tests will be added.

This commit use `ENVIRONMENT_MODIFICATION` instead of `ENVIRONMENT`, since adding a value to the PATH in windows is a mess otherwise with CMake, because `;` is used as a list separator.

I took the liberty to refactor some boilerplate code, to make it working
more easily. More refactoring is needed to remove boilerplate code.